code-review-api/requirements.txt
# Web框架
fastapi==0.115.10
uvicorn[standard]==0.22.0

# 数据库
sqlalchemy==2.0.39
pymysql==1.1.0

# 安全认证
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 工具
python-dotenv==1.0.0
pydantic==2.10.6  
email-validator==2.0.0
pytz==2023.3
pyyaml==6.0.1
psutil==6.0.0
# 开发工具
pylint==2.17.2
flake8==6.0.0

# 如果使用异步任务
celery==5.3.4
redis==4.5.5

# 指定镜像源
--index-url https://mirrors.aliyun.com/pypi/simple/
